Create single fixed user for program use in MySQL
We switched the SQL database so that we all use user politronix@localhost to write to the DB. That way we don't have to edit the code for each person, and it won't switch users every time you pull/push. Plus, we'll need a user in prod use on AWS, so it might as well work this way.